Find cell clusters in 3D spatial data using alpha hull clustering algorithm.

This function finds cell clusters in a 3D SpatialExperiment object using the alpha hull clustering algorithm.

alpha_hull_clustering3D(
  spe,
  cell_types_of_interest,
  alpha,
  minimum_cells_in_cluster,
  feature_colname,
  plot_image = TRUE
)

Arguments

spe

A SpatialExperiment object containing 3D spatial information for the cells. Naming of spatial coordinates MUST be "Cell.X.Position", "Cell.Y.Position", "Cell.Z.Position" for the x-coordinate, y-coordinate and z-coordinate of each cell.

cell_types_of_interest

A character vector specifying the cell types of interest. Clusters containing only these cells will be found.

alpha

A positive numeric. A smaller alpha value results in clusters with more intricate borders. A large alpha value results in clusters with smooth and simple boundaries.

minimum_cells_in_cluster

A positive numeric. Clusters identified which have less than this specified value are relabelled as not a cluster.

feature_colname

A string specifying the name of the column in the colData slot of the SpatialExperiment object that contains the cell type information.

plot_image

A logical indicating whether to plot 3D spatial data with alpha hull clusters. Defaults to TRUE.

Value

The same 3D SpatialExperiment object used as input for spe, with an added column in the colData slot to specify which alpha hull cluster each cell belongs to, and added metadata containing information needed to plot the alpha hull clusters.

Examples

# Get simulated SpatialExperiment object to use as an example for analysis
simulated_spe <- readRDS(system.file("extdata", "simulated_spe.rds", package = "SPIAT3D"))
#> Loading required namespace: SpatialExperiment

# Alpha hull clustering depends on alphashape3d package, which depends on rgl
# package, and you might need to run this code, as newer mac versions don't
# support rgl.
if (Sys.info()[["sysname"]] == "Darwin") {
  options(rgl.useNULL = TRUE)
  options(rgl.printRglwidget = TRUE)
}

# Alpha hull clustering
alpha_hull_spe <- alpha_hull_clustering3D(
    spe = simulated_spe,
    cell_types_of_interest = c("Tumour", "Immune"),
    alpha = 8,
    minimum_cells_in_cluster = 5,
    feature_colname = "Cell.Type",
    plot_image = TRUE
)